backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
mysql-connector-python
PyJWT
bcrypt
python-dotenv
requests
numpy>=1.26.0
pandas
scikit-learn
tensorflow
Pillow
opencv-python-headless
werkzeug
user-agents
geoip2
psutil
